// client/src/old_stuff/types/general.ts
export enum SyncStatus {
  Cancelled = 'cancelled',
  Cancelling = 'cancelling',
  Uninitialized = 'uninitialized',
  Queued = 'queued',
  Done = 'done',
  Error = 'error',
  Removed = 'removed',
  Indexing = 'indexing',
  Syncing = 'syncing',
  RemoteRemoved = 'remote_removed',
}

export enum RepoProvider {
  GitHub = 'github',
  Local = 'local',
}

export type RepoType = {
  ref: string;
  name: string;
  provider: RepoProvider;
  local_duplicates: string[];
  last_update: string;
  last_index: string;
  sync_status: SyncStatus;
  most_common_lang: string;
  branches: { name: string; last_commit_unix_secs: number }[];
  branch_filter: { select: string[] } | null;
};

export type RepoUi = RepoType & {
  shortName: string;
  folderName: string;
  alreadySynced?: boolean;
};

export type CodeStudioShortType = {
  id: string;
  name: string;
  modified_at: string;
  repos: string[];
  most_common_ext: string;
};

export enum FullResultModeEnum {
  PAGE,
  SIDEBAR,
  MODAL,
}

export enum SearchType {
  REGEX,
  NL,
}

export enum TabType {
  REPO = 'repo',
  STUDIO = 'studio',
  HOME = 'home',
}

export type RepoTabType = {
  key: string;
  name: string;
  repoName: string;
  repoRef: string;
  source: RepoSource;
  branch?: string | null;
  navigationHistory: NavigationItem[];
  type: TabType.REPO;
};

export type HomeTabType = {
  key: string;
  name: string;
  type: TabType.HOME;
};

export type StudioTabType = {
  key: string;
  name: string;
  type: TabType.STUDIO;
};

export type UITabType = RepoTabType | HomeTabType | StudioTabType;
// ...
export interface NavigationItem {
  type:
    | 'search'
    | 'repo'
    | 'full-result'
    | 'home'
    | 'conversation-result'
    | 'article-response';
  query?: string;
  repo?: string;
  path?: string;
  page?: number;
  loaded?: boolean;
  isInitial?: boolean;
  searchType?: SearchType;
  pathParams?: Record<string, string>;
  threadId?: string;
  recordId?: number;
}